requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
opencv-python>=4.5.0
matplotlib>=3.5.0
scipy>=1.7.0
tqdm>=4.62.0
Pillow>=8.3.0
tensorboard>=2.8.0
wandb>=0.13.0
plyfile>=0.7.4
trimesh>=3.9.0
open3d>=0.15.0
einops>=0.6.0
transformers>=4.20.0
accelerate>=0.20.0
